fix first problem with delete, update some CLI function args in review class
