+++ config/environment.rb
@@ -1,7 +1,6 @@
 require 'bundler'
 require 'pry'
 require 'json'
-require 'unirest'
 require 'rake'
 require 'sqlite3'
