+++ mera/video_flow.md
@@ -1,0 +1,16 @@
+
+
+A nice order to show our functionality in our video
+
+- see reviews
+  (no reviews, tells you to go search)
+- basic search (cuisine)
+- choose restaurant
+- go back
+- choose another restaurant
+- create a review of the restaurant
+- see reviews
+- update that review
+- see reviews
+- delete it
+- go try out the more complex search
